Elements: Push as Mixin Class
Move the particle container push logic over to the elements itself.
